Added events for WindowsWindow and handled WindowClose

+++ Dough/Include/Dough/Events/MouseEvent.h
@@ -7,15 +7,15 @@ namespace Dough
 	class DOUGH_API MouseMovedEvent : public Event
 	{
 	public:
-		MouseMovedEvent(int x, int y)
+		MouseMovedEvent(float x, float y)
 			: m_MouseX(x), m_MouseY(y)
 		{}
 
 		EVENT_CLASS_CATEGORY(EC_Mouse | EC_Input)
 		EVENT_CLASS_TYPE(MouseMoved)
 
-		[[nodiscard]] inline int GetX() const { return m_MouseX; }
-		[[nodiscard]] inline int GetY() const { return m_MouseY; }
+		[[nodiscard]] inline float GetX() const { return m_MouseX; }
+		[[nodiscard]] inline float GetY() const { return m_MouseY; }
 
 		[[nodiscard]] std::string ToString() const override
 		{
@@ -25,21 +25,21 @@ namespace Dough
 		}
 
 	private:
-		int m_MouseX, m_MouseY;
+		float m_MouseX, m_MouseY;
 	};
 
 	class DOUGH_API MouseScrolledEvent : public Event
 	{
 	public:
-		MouseScrolledEvent(int xOffset, int yOffset)
+		MouseScrolledEvent(double xOffset, double yOffset)
 			: m_XOffset(xOffset), m_YOffset(yOffset)
 		{}
 
 		EVENT_CLASS_CATEGORY(EC_Mouse | EC_Input)
 		EVENT_CLASS_TYPE(MouseScrolled)
 
-		[[nodiscard]] inline int GetXOffset() const { return m_XOffset; }
-		[[nodiscard]] inline int GetYOffset() const { return m_YOffset; }
+		[[nodiscard]] inline double GetXOffset() const { return m_XOffset; }
+		[[nodiscard]] inline double GetYOffset() const { return m_YOffset; }
 
 		[[nodiscard]] std::string ToString() const override
 		{
@@ -49,7 +49,7 @@ namespace Dough
 		}
 
 	private:
-		int m_XOffset, m_YOffset;
+		double m_XOffset, m_YOffset;
 	};
 
 	class DOUGH_API MouseButtonPressedEvent : public Event
